+++ src/core/hle/ipc_helpers.h
@@ -423,7 +423,7 @@ inline const std::vector<u8>& RequestParser::PopStaticBuffer() {
     Pop<VAddr>();
 
     StaticBufferDescInfo buffer_info{sbuffer_descriptor};
-    return context->GetStaticBuffer(buffer_info.buffer_id);
+    return context->GetStaticBuffer(static_cast<u8>(buffer_info.buffer_id));
 }
 
 inline Kernel::MappedBuffer& RequestParser::PopMappedBuffer() {
